Minor fixes: a few more ulongs, CONTRIBUTORS, magic numbers

+++ sys/src/boot/pc/lib.h
@@ -7,11 +7,11 @@
 /*
  * mem routines
  */
-extern	void*	memccpy(void*, void*, int, ulong);
+extern	void*	memccpy(void*, void*, int, usize);
 extern	void*	memset(void*, int, usize);
 extern	int	memcmp(void*, void*, usize);
 extern	void*	memmove(void*, void*, usize);
-extern	void*	memchr(void*, int, ulong);
+extern	void*	memchr(void*, int, usize);
 
 /*
  * string routines
